+++ VSC Extension/src/config/website.js
@@ -1,6 +1,6 @@
 module.exports = `
 <!DOCTYPE html>
-<link rel="icon" href="https://raw.githubusercontent.com/DervexHero/Argon/main/VSC%20Extension/images/Logo.png">
+<link rel="icon" href="https://raw.githubusercontent.com/DervexHero/Argon/main/VSC%20Extension/assets/Logo.png">
 
 <html>
 
@@ -14,7 +14,7 @@ h1 {color: rgb(240, 240, 240); font-family: Roboto Mono; font-weight: normal;}
 </style>
 
 <body>
-<img src="https://raw.githubusercontent.com/DervexHero/Argon/main/VSC%20Extension/images/LogoName.png" alt="Argon logo and name">
+<img src="https://raw.githubusercontent.com/DervexHero/Argon/main/VSC%20Extension/assets/LogoName.png" alt="Argon logo and name">
 <h1>Server uptime: <b>$time</b></h1>
 <h1>Sync count: <b>$synces</b></h1>
 </body>